Add matrix multiplication

+++ cpp_samples/matrixVectorMul.cpp
@@ -21,12 +21,12 @@ int main(int argc, char **argv)
     std::generate(x.begin(), x.end(), [](){return rand() / (float)RAND_MAX;});
     std::generate(y.begin(), y.end(), [](){return rand() / (float)RAND_MAX;});
 
-    for(int row = 0; row < rows; ++row)
+    for(unsigned row = 0; row < rows; ++row)
     {
-        double sum = 0.0;
-        for(int col = 0; col < cols; ++col)
+        float sum = 0.0;
+        for(unsigned col = 0; col < cols; ++col)
             sum += x[row * cols + col] * y[col];
-        result[row] = static_cast<float>(sum);
+        result[row] = sum;
     }
 
     matrixVectorMul(x.data(), y.data(), cudaResult.data(), rows, cols);
